get pre-commit back working with imports
